a few tests fixed/rewritten
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@268 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ tests/cxts/cxts.cpp
@@ -1120,8 +1120,8 @@ int CvTS::find_written_param( CvTest* test, const char* paramname, int valtype, 
     bool add_to_list = test->get_func_list()[0] == '\0';
     char buffer[256];
     int paramname_len = (int)strlen(paramname);
-    int paramval_len = valtype == CV_NODE_INT ? sizeof(int) :
-        valtype == CV_NODE_REAL ? sizeof(double) : -1;
+    int paramval_len = valtype == CV_NODE_INT ? (int)sizeof(int) :
+        valtype == CV_NODE_REAL ? (int)sizeof(double) : -1;
     const char* name = CvTest::get_parent_name( testname, buffer );
 
     if( !fs )
